requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Session==0.4.0
numpy==1.25.2
pandas==2.1.1
scikit-learn==1.3.1
mysql-connector-python==8.1.0
gunicorn==20.1.0
jupyter==1.0.0
nbconvert==7.7.4
